在idt.c文件中把int 写为了char导致了死循环，现在出现了新的bug，就是只出现中断2

+++ src/test3/idt/idt.c
@@ -42,7 +42,7 @@ void init_idt()
 {
 	bzer((uint8*)&interrupt_handlers, sizeof(interrupt_handler_t) * 256);
 	idt_ptr.limit = sizeof(idt_entry_t) * 256 - 1;
-	idt_ptr.base = (uint8)&idt_entries;
+	idt_ptr.base = (uint32)&idt_entries;
 
 	bzer((uint8 *)&idt_entries, sizeof(idt_entry_t) * 256);
